+++ code_examples/learning_by_doing/section1_samples/type_error.py
