+++ reference/code_examples/learning_by_doing/section1_samples/index.html
@@ -26,13 +26,18 @@
 <section>
 <h2 class="section-title" id="header-submodules">Sub-modules</h2>
 <dl>
-<dt><code class="name"><a title="code_examples.learning_by_doing.section1_samples.exercise4_type_error" href="exercise4_type_error.html">code_examples.learning_by_doing.section1_samples.exercise4_type_error</a></code></dt>
-<dd>
-<div class="desc"><p>Shows an example of type error.</p></div>
-</dd>
 <dt><code class="name"><a title="code_examples.learning_by_doing.section1_samples.main" href="main.html">code_examples.learning_by_doing.section1_samples.main</a></code></dt>
 <dd>
-<div class="desc"><p>Executes samples that show the use of basic Python syntax …</p></div>
+<div class="desc"><p>Executes samples that show the use of basic Python syntax.
+Adapted from [How to create a menu for a python console …</p></div>
+</dd>
+<dt><code class="name"><a title="code_examples.learning_by_doing.section1_samples.sequence_indexing" href="sequence_indexing.html">code_examples.learning_by_doing.section1_samples.sequence_indexing</a></code></dt>
+<dd>
+<div class="desc"><p>Example of sequence indexing.</p></div>
+</dd>
+<dt><code class="name"><a title="code_examples.learning_by_doing.section1_samples.type_error" href="type_error.html">code_examples.learning_by_doing.section1_samples.type_error</a></code></dt>
+<dd>
+<div class="desc"><p>Shows an example of type error.</p></div>
 </dd>
 </dl>
 </section>
@@ -56,8 +61,9 @@
 </li>
 <li><h3><a href="#header-submodules">Sub-modules</a></h3>
 <ul>
-<li><code><a title="code_examples.learning_by_doing.section1_samples.exercise4_type_error" href="exercise4_type_error.html">code_examples.learning_by_doing.section1_samples.exercise4_type_error</a></code></li>
 <li><code><a title="code_examples.learning_by_doing.section1_samples.main" href="main.html">code_examples.learning_by_doing.section1_samples.main</a></code></li>
+<li><code><a title="code_examples.learning_by_doing.section1_samples.sequence_indexing" href="sequence_indexing.html">code_examples.learning_by_doing.section1_samples.sequence_indexing</a></code></li>
+<li><code><a title="code_examples.learning_by_doing.section1_samples.type_error" href="type_error.html">code_examples.learning_by_doing.section1_samples.type_error</a></code></li>
 </ul>
 </li>
 </ul>
